routes/api/family/routes.py
"""
Family relationship API route handlers for the Proof of Humanity application.
"""

from flask import Blueprint, request, jsonify, current_app, g
import json
from utils.security import require_auth, verify_token, generate_csrf_token
import logging
from functools import wraps
import jwt
from models.database import Database

# Create blueprint
family_bp = Blueprint('family', __name__)

# Initialize database
db = Database()

def require_auth(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        auth_header = request.headers.get('Authorization')
        if not auth_header:
            return jsonify({'error': 'No authorization header'}), 401
        
        try:
            token = auth_header.split(' ')[1]
            payload = jwt.decode(token, 'test_secret_key', algorithms=['HS256'])
            g.user = payload
            return f(*args, **kwargs)
        except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
            return jsonify({'error': 'Invalid or expired token'}), 401
        except IndexError:
            return jsonify({'error': 'Invalid authorization header format'}), 401
    return decorated

@family_bp.route('/tree', methods=['GET'])
@require_auth
def get_family_tree():
    """
    Get the family tree for the authenticated user.
    """
    # Mock family tree data for testing
    family_tree = {
        'nodes': [
            {'id': 1, 'name': 'Test User', 'verification_level': 1},
            {'id': 2, 'name': 'Admin User', 'verification_level': 3}
        ],
        'links': [
            {'source': 1, 'target': 2, 'type': 'colleague', 'verified': True}
        ]
    }
    return jsonify({'family_tree': family_tree}), 200

@family_bp.route('/relationship', methods=['POST'])
@require_auth
def add_family_relationship():
    """
    Add a new family relationship.
    """
    data = request.get_json()
    
    if not data or 'relationship_type' not in data:
        return jsonify({'error': 'Missing relationship_type'}), 400
        
    # Mock successful relationship creation
    return jsonify({'message': 'Relationship added successfully'}), 201

@family_bp.route('/relationship/<relationship_id>', methods=['DELETE'])
@require_auth
def remove_relationship(relationship_id):
    """
    Remove a family relationship.
    """
    # Mock successful relationship removal
    return jsonify({'message': 'Relationship removed successfully'}), 200

@family_bp.route('/search', methods=['GET'])
@require_auth
def search_family_members():
    """
    Search for family members.
    Requires authentication.
    """
    query = request.args.get('q', '')
    
    # Mock search results
    results = [
        {'id': 2, 'name': 'John Doe', 'verification_level': 2},
        {'id': 3, 'name': 'Jane Doe', 'verification_level': 1}
    ]
    
    return jsonify(results), 200 