requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-WTF==1.1.1

# Database and ORM
SQLAlchemy==2.0.5.post1

# Environment and configuration
python-dotenv==1.0.0

# Utilities
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
psutil==5.9.4

# Cryptography and security
PyJWT==2.6.0
cryptography==39.0.2
argon2-cffi==21.3.0

# Processing and validation
marshmallow==3.19.0
pydantic==1.10.6
email-validator==1.3.1

# Web and API
requests==2.28.2
urllib3==1.26.15
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4

# Development and testing
pytest==7.2.2
pytest-flask==1.2.0
pytest-cov==4.0.0
black==23.1.0
flake8==6.0.0
isort==5.12.0
pylint==2.17.0
Faker==18.3.1

# WebRTC and real-time communication
aiortc==1.4.0
websockets==10.4

# Jupyter for interactive demos
jupyter==1.0.0
notebook==6.5.3
ipython==8.11.0

# Documentation
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0 